learning/requirements.txt
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
importlib-metadata==1.7.0
six==1.15.0
virtualenv==20.0.25
zipp==3.1.0
torch==1.5.1
rasterio==1.1.5
pandas==1.0.5
numpy==1.19.1
pillow==7.2.0
pyyaml==5.3.1
torchvision==0.6.1
addict==2.2.1
